+++ sources/orthanc_auth_service/app.py
@@ -15,11 +15,23 @@ import pytz
 from shares.models import *
 from shares.orthanc_token_service_factory import create_token_service_from_secrets
 from shares.keycloak import create_keycloak_from_secrets
+from shares.roles_configuration import RolesConfiguration, create_roles_configuration_from_file
+from shares.api_keys import create_api_keys_from_file
 
 logging.basicConfig(level=logging.DEBUG)
 
 token_service = create_token_service_from_secrets()
-keycloak = create_keycloak_from_secrets()
+
+permissions_file_path = os.environ.get("PERMISSIONS_FILE_PATH", "/orthanc_auth_service/permissions.json")
+api_keys_file_path = os.environ.get("API_KEYS_FILE_PATH", "/orthanc_auth_service/api-keys.json")
+
+roles_configuration = create_roles_configuration_from_file(permissions_file_path)
+keycloak = create_keycloak_from_secrets(roles_configuration)
+
+api_keys = None
+if roles_configuration and api_keys_file_path:
+    api_keys = create_api_keys_from_file(api_keys_file_path=api_keys_file_path,
+                                         roles_configuration=roles_configuration)
 app = FastAPI()
 
 # check if the service requires basic auth (by checking of some USERS have been defined)
@@ -167,8 +179,13 @@ def get_user_profile(user_profile_request: UserProfileRequest):
         if keycloak is None:
             logging.warning("Keycloak is not configured, all users are considered anonymous")
             return anonymous_profile
+
         elif user_profile_request.token_key is not None:
-            response = keycloak.get_user_profile_from_token(user_profile_request.token_value)
+            if user_profile_request.token_key == "api-key" and api_keys is not None:
+                response = api_keys.get_user_profile_from_api_key(api_key=user_profile_request.token_value)
+            else:
+                response = keycloak.get_user_profile_from_token(user_profile_request.token_value)
+
         else:
             return anonymous_profile